% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{ggplot_tox_probas}
\alias{ggplot_tox_probas}
\title{Graph of prior and posterior toxicity probabilities}
\usage{
ggplot_tox_probas(
  modelfit,
  dose_level = "dose",
  tox_probas = "mean_prob_tox",
  skeleton = "Skeleton",
  target = 0.25
)
}
\arguments{
\item{target}{}
}
\description{
Graph of prior and posterior toxicity probabilities
}
