% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{tox_df_to_vector}
\alias{tox_df_to_vector}
\title{Formatting history of patient toxicities into a vector suited for escalation::fit}
\usage{
tox_df_to_vector(data, dose_level, tox_probas)
}
\arguments{
\item{tox_probas}{}
}
\description{
Formatting history of patient toxicities into a vector suited for escalation::fit
}
