% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{ggplot_tox_probas}
\alias{ggplot_tox_probas}
\title{Graph of prior and posterior toxicity probabilities}
\usage{
ggplot_tox_probas(modelfit, target = 0.25, CI = F)
}
\arguments{
\item{target}{}
}
\description{
Graph of prior and posterior toxicity probabilities
}
